Initial guide write & reorganization - Added an '.editorconfig' for enforcing style guides and GitHub file formatting - Updated '.gitignore' formatting and added `.DS_Store` for macOS devices - Changed `groupId` and `artifactId` throughout (moving away from 'GDI_Hackathon' references and `groupId` now matches newly aquired domain: 'shybye.app') - Removed `.vscode/` from '.gitignore' to add workspace extension suggestions and settings - Migrated all content ('ShyBye/') to root directory, project and startup can now be initialized from the root directory - [bugfix] addressed an error associated with `maven-surefire-plugin` see: https://stackoverflow.com/a/36429564 - Added proper description to 'pom.xml' - Formatted all files
