+++ .editorconfig
@@ -1,0 +1,18 @@
+# Learn more: https://EditorConfig.org
+
+
+root = true
+
+[*]
+# Adheres to Maven Apache style guide
+# Learn more: https://maven.apache.org/guides/development/guide-documentation-style
+indent_style = space
+indent_size = 2
+end_of_line = lf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.java]
+indent_style = tab
+indent_size = 4
