+++ .gitignore
@@ -1,0 +1,33 @@
+# Learn more: https://help.github.com/articles/ignoring-files
+
+HELP.md
+target/
+.DS_Store
+!.mvn/wrapper/maven-wrapper.jar
+!**/src/main/**/target/
+!**/src/test/**/target/
+
+# STS
+.apt_generated
+.classpath
+.factorypath
+.project
+.settings
+.springBeans
+.sts4-cache
+
+# IntelliJ IDEA
+.idea
+*.iml
+*.ipr
+*.iws
+
+# NetBeans
+build/
+/dist/
+/nbbuild/
+/nbdist/
+/nbproject/private/
+/.nb-gradle/
+!**/src/main/**/build/
+!**/src/test/**/build/
