+++ .vscode/extensions.json
@@ -1,0 +1,6 @@
+{
+	"recommendations": [
+		"vscjava.vscode-java-pack",
+		"redhat.vscode-xml"
+	]
+}
