+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+	"java.compile.nullAnalysis.mode": "automatic",
+  "[java]": {
+		"editor.defaultFormatter": "redhat.java"
+	},
+  "[xml]": {
+		"editor.defaultFormatter": "redhat.vscode-xml"
+	}
+}
