--- ShyBye/src/main/java/com/GDI_Hackathon/ShyBye/Models/User.java
@@ -1,67 +1,0 @@
-package com.GDI_Hackathon.ShyBye.Models;
-
-public class User {
-
-    private int userId;
-
-    private  String username;
-
-    private  String password;
-
-    private int userPoints = 0 ;
-
-    public User(int userId, String username, String password){
-        this.userId = userId;
-        this.username = username;
-        this.password = password;
-    }
-
-    public User() {
-    }
-
-    public int getUserId() {
-        return userId;
-    }
-
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public int getUserPoints() {
-        return userPoints;
-    }
-
-    public void setUserPoints(int userPoints) {
-        this.userPoints = userPoints;
-    }
-
-    public void addToScore(int points){
-        this.userPoints += points;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "userId=" + userId +
-                ", username='" + username + '\'' +
-                ", password='" + password + '\'' +
-                ", userPoints=" + userPoints +
-                '}';
-    }
-}
