--- ShyBye/src/main/java/com/GDI_Hackathon/ShyBye/ShyByeApplication.java
@@ -1,13 +1,0 @@
-package com.GDI_Hackathon.ShyBye;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShyByeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ShyByeApplication.class, args);
-	}
-
-}
