--- ShyBye/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-#datasource connection variables
-#spring.datasource.url=jdbc:postgresql://localhost:5432/shy_byedb
-spring.datasource.url=jdbc:postgresql://shy-bye-db.internal:5432/shybyedb
-spring.datasource.name=shybyedb
-spring.datasource.username=postgres
-spring.datasource.password=postgres1
-
-#Server settings
-server.error.include.stacktrace=never
-server.port=9000
