--- ShyBye/src/test/java/com/GDI_Hackathon/ShyBye/ShyByeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.GDI_Hackathon.ShyBye;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShyByeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
