+++ src/main/java/app/shybye/exceptions/UserIdNotFoundException.java
@@ -1,0 +1,8 @@
+package app.shybye.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "UserId not found")
+public class UserIdNotFoundException extends RuntimeException {
+}
