+++ src/main/java/app/shybye/models/LoginDTO.java
@@ -1,0 +1,39 @@
+package app.shybye.models;
+
+public class LoginDTO {
+
+	private String username;
+	private String password;
+
+	public LoginDTO(String username, String password) {
+		this.username = username;
+		this.password = password;
+	}
+
+	public LoginDTO() {
+	}
+
+	public String getUsername() {
+		return username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	@Override
+	public String toString() {
+		return "LoginDTO{" +
+				"username='" + username + '\'' +
+				", passwrd='" + password + '\'' +
+				'}';
+	}
+}
