+++ src/test/java/app/shybye/ApplicationTests.java
@@ -1,0 +1,13 @@
+package app.shybye;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
